// src/enums/categories.enum.ts
export enum CategoriesEnum {
  ALIMENTATION = "Alimentation",
  SHOPPING = "Shopping et services",
  LOGEMENT = "Logement - maison",
  A_CATEGORISER_RENTREE = "A categoriser - rentree d'argent",
  A_CATEGORISER_SORTIE = "A categoriser - sortie d'argent",
  SANTE = "Sante",
  BANQUES_ASSURANCES = "Banque et assurances",
  LOISIRS = "Loisirs et vacances",
  REVENUS = "Revenus et rentrees d'argent",
  EDUCATION = "Education et famille",
}